Revert "Revert "MAGN-5604 Remove execution throttle.""
