+++ src/DynamoCore/Models/DynamoModelEvents.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Windows;
 
 using Dynamo.Core;
