+++ src/DynamoCore/Models/NodeModel.cs
@@ -23,7 +23,7 @@ using ProtoCore.DSASM;
 
 namespace Dynamo.Models
 {
-    public abstract class NodeModel : ModelBase, IBlockingModel
+    public abstract class NodeModel : ModelBase
     {
         #region private members
