+++ src/DynamoCore/Utilities/CodeBlockUtils.cs
@@ -1,8 +1,4 @@
-﻿using System.Globalization;
-using System.Windows;
-using System.Windows.Media;
-
-using Dynamo.UI;
+﻿using Dynamo.UI;
 
 using Dynamo.Models;
 using Dynamo.Nodes;
