+++ src/DynamoCoreWpf/Controls/DynamoTextBox.cs
@@ -150,6 +150,11 @@ namespace Dynamo.Nodes
         new public string Text
         {
             get { return base.Text; }
+            set
+            {
+                base.Text = value;
+                UpdateDataSource(true);
+            }
         }
 
         #endregion
@@ -258,39 +263,6 @@ namespace Dynamo.Nodes
 
         #endregion
     }
-
-    public class DynamoSlider : Slider
-    {
-        NodeModel nodeModel;
-        public DynamoSlider(NodeModel model)
-        {
-            nodeModel = model;
-        }
-
-        #region Event Handlers
-        protected override void OnThumbDragStarted(System.Windows.Controls.Primitives.DragStartedEventArgs e)
-        {
-            base.OnThumbDragStarted(e);
-            nodeModel.Workspace.RecordModelForModification(nodeModel);
-            (nodeModel as IBlockingModel).OnBlockingStarted(EventArgs.Empty);
-        }
-
-        protected override void OnThumbDragCompleted(System.Windows.Controls.Primitives.DragCompletedEventArgs e)
-        {
-            base.OnThumbDragCompleted(e);
-            (nodeModel as IBlockingModel).OnBlockingEnded(EventArgs.Empty);
-            nodeModel.RequiresRecalc = true;
-        }
-
-        protected override void OnPreviewMouseLeftButtonDown(MouseButtonEventArgs e)
-        {
-            base.OnPreviewMouseLeftButtonDown(e);
-            if (e.OriginalSource is System.Windows.Shapes.Rectangle)
-                nodeModel.Workspace.RecordModelForModification(nodeModel);
-        }
-        #endregion
-    }
-
 }
 
 namespace Dynamo.UI.Controls
