+++ src/Libraries/CoreNodesUI/Input/SliderBase.cs
@@ -2,141 +2,10 @@
 using System.Globalization;
 using System.Xml;
 
-using Dynamo.Controls;
 using Dynamo.Models;
 
-using Autodesk.DesignScript.Runtime;
-
-using Microsoft.Practices.Prism.ViewModel;
-
-namespace Dynamo.Nodes
+namespace DSCoreNodesUI.Input
 {
-    public enum NumericFormat{Double, Integer}
-
-    /// <summary>
-    /// The SliderViewModel acts as the converter
-    /// for numeric sliders. By using a view model
-    /// to do the conversion instead of a converter,
-    /// we can do conditional conversion based on the 
-    /// context in which the conversion happens.
-    /// </summary>
-    public class SliderViewModel: NotificationObject
-    {
-        private NumericFormat format;
-        private SliderBase model;
-
-        public string MaxText
-        {
-            get { return ConvertToString(format, model.Max); }
-        }
-
-        public string MinText
-        {
-            get{ return ConvertToString(format, model.Min); }
-        }
-
-        public string StepText
-        {
-            get{return ConvertToString(format, model.Step); }
-        }
-
-        public string ValueText
-        {
-            get { return ConvertToString(format, model.Value); }
-        }
-
-        public double Max
-        {
-            get { return model.Max; }
-        }
-
-        public double Min
-        {
-            get { return model.Min; }
-        }
-
-        public double Step
-        {
-            get { return model.Step; }
-        }
-
-        public double Value
-        {
-            get { return model.Value; }
-            set
-            {
-                if (value >= model.Max)
-                    model.Max = value;
-
-                if (value <= model.Min)
-                    model.Min = value;
-
-                model.Value = value;
-            }
-        }
-
-        public SliderViewModel(NumericFormat format, SliderBase sliderBaseModel)
-        {
-            this.format = format;
-            model = sliderBaseModel;
-            model.PropertyChanged += model_PropertyChanged;
-        }
-
-        void model_PropertyChanged(object sender, System.ComponentModel.PropertyChangedEventArgs e)
-        {
-            switch (e.PropertyName)
-            {
-                case "Max":
-                    RaisePropertyChanged("Max");
-                    RaisePropertyChanged("MaxText");
-                    break;
-                case "Min":
-                    RaisePropertyChanged("Min");
-                    RaisePropertyChanged("MinText");
-                    break;
-                case "Value":
-                    RaisePropertyChanged("Value");
-                    RaisePropertyChanged("ValueText");
-                    break;
-                case "Step":
-                    RaisePropertyChanged("Step");
-                    RaisePropertyChanged("StepText");
-                    break;
-            }
-        }
-
-        internal static string ConvertToString(NumericFormat format, double value)
-        {
-            switch (format)
-            {
-                case NumericFormat.Double:
-                    return value.ToString("0.000", CultureInfo.InvariantCulture);
-                case NumericFormat.Integer:
-                    return value.ToString("0", CultureInfo.InvariantCulture);
-            }
-
-            return "0.0";
-        }
-
-        internal static double ConvertToDouble(NumericFormat format, string value)
-        {
-            switch (format)
-            {
-                case NumericFormat.Double:
-                    double d = 0.0;
-                    double.TryParse(value, NumberStyles.Any, CultureInfo.InvariantCulture, out d);
-                    return d;
-
-                case NumericFormat.Integer:
-                    int i = 0;
-                    int.TryParse(value, NumberStyles.Any, CultureInfo.InvariantCulture, out i);
-                    return i;
-            }
-
-            return 0.0;
-        }
-    }
-
     /// <summary>
     /// SliderBase contains logic for range and 
     /// step values for slider classes.
@@ -180,7 +49,12 @@ namespace Dynamo.Nodes
             }
         }
 
-        protected SliderBase(WorkspaceModel workspace): base(workspace){}
+        protected override bool ShouldDisplayPreviewCore
+        {
+            get { return false; }
+        }
+
+        protected SliderBase(WorkspaceModel workspace) : base(workspace){}
 
         protected override bool UpdateValueCore(string name, string value)
         {
@@ -213,11 +87,6 @@ namespace Dynamo.Nodes
             }
 
             return base.UpdateValueCore(name, value);
-        }
-
-        protected override bool ShouldDisplayPreviewCore()
-        {
-            return false; // Previews are not shown for this node type.
         }
 
         #region Load/Save
@@ -325,68 +194,4 @@ namespace Dynamo.Nodes
         #endregion
     }
 
-    [NodeName("Number Slider")]
-    [NodeCategory(BuiltinNodeCategories.CORE_INPUT)]
-    [NodeDescription("A slider that produces numeric values.")]
-    [SupressImportIntoVM]
-    [IsDesignScriptCompatible]
-    [NodeSearchTags(new []{"double","number","float","integer","slider"})]
-    public class DoubleSlider : SliderBase
-    {
-        public DoubleSlider(WorkspaceModel workspace) : base(workspace)
-        {
-            Min = 0;
-            Max = 100;
-            Step = 0.01;
-            Value = 0;
-        }
-
-        /// <summary>
-        /// UI is initialized and bindings are setup here.
-        /// </summary>
-        /// <param name="nodeUI">UI view that we can customize the UI of.</param>
-        public override void SetupCustomUIElements(dynNodeView nodeUI)
-        {
-            var slider = new UI.Controls.DynamoSlider(this, nodeUI)
-            {
-                DataContext = new SliderViewModel(NumericFormat.Double, this)
-            };
-
-            nodeUI.inputGrid.Children.Add(slider);
-        }
-
-    }
-
-    [NodeName("Integer Slider")]
-    [NodeCategory(BuiltinNodeCategories.CORE_INPUT)]
-    [NodeDescription("A slider that produces integer values.")]
-    [SupressImportIntoVM]
-    [IsDesignScriptCompatible]
-    public class IntegerSlider : SliderBase
-    {
-        public IntegerSlider(WorkspaceModel workspace) : base(workspace)
-        {
-            RegisterAllPorts();
-
-            Min = 0;
-            Max = 100;
-            Step = 1;
-            Value = 0;
-        }
-
-        /// <summary>
-        /// UI is initialized and bindings are setup here.
-        /// </summary>
-        /// <param name="nodeUI">UI view that we can customize the UI of.</param>
-        public override void SetupCustomUIElements(dynNodeView nodeUI)
-        {
-            var slider = new UI.Controls.DynamoSlider(this, nodeUI)
-            {
-                DataContext = new SliderViewModel(NumericFormat.Integer,this)
-            };
-
-            nodeUI.inputGrid.Children.Add(slider);
-        }
-
-    }
 }
