+++ src/Libraries/CoreNodesUI/UI/DynamoSlider.xaml
@@ -1,0 +1,83 @@
+﻿<UserControl x:Class="Dynamo.UI.Controls.DynamoSlider"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:ui1="clr-namespace:Dynamo.UI;assembly=DynamoCoreWpf"
+             xmlns:ui2="clr-namespace:Dynamo.UI;assembly=DynamoCore"
+             xmlns:nodes1="clr-namespace:Dynamo.Nodes;assembly=DynamoCoreWpf"
+             mc:Ignorable="d" 
+             HorizontalAlignment="Stretch"
+             VerticalAlignment="Stretch">
+    
+    <UserControl.Resources>
+        <ResourceDictionary>
+            <ResourceDictionary.MergedDictionaries>
+                <ui1:SharedResourceDictionary Source="{x:Static ui1:SharedDictionaryManager.DynamoConvertersDictionaryUri}"/>
+            </ResourceDictionary.MergedDictionaries>
+        </ResourceDictionary>
+    </UserControl.Resources>
+    
+    <Grid VerticalAlignment="Stretch">
+ 
+        <Expander IsExpanded="False" Margin="5,0,0,0" IsHitTestVisible="True">
+
+            <Grid Margin="0,7,0,0">
+                
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="40"/>
+                    <ColumnDefinition Width="*"/>
+                </Grid.ColumnDefinitions>
+                
+                <Grid.RowDefinitions>
+                    <RowDefinition Height="30"/>
+                    <RowDefinition Height="30"/>
+                    <RowDefinition Height="30"/>
+                    <RowDefinition Height="Auto"/>
+                </Grid.RowDefinitions>
+
+                <Label Grid.Column="0" Grid.Row="0" Content="Min" VerticalContentAlignment="Center"/>
+                <Label Grid.Column="0" Grid.Row="1" Content="Max" VerticalContentAlignment="Center"/>
+                <Label Grid.Column="0" Grid.Row="2" Content="Step" VerticalContentAlignment="Center"/>
+                <nodes1:DynamoTextBox x:Name="MinTb" Grid.Row="0" Grid.Column="1" Margin="2"
+                                        VerticalContentAlignment="Center" 
+                                        Text="{Binding MinText, Mode=OneWay, UpdateSourceTrigger=Explicit}"/>
+                <nodes1:DynamoTextBox x:Name="MaxTb" Grid.Row="1" Grid.Column="1" Margin="2"
+                                        VerticalContentAlignment="Center" 
+                                        Text="{Binding MaxText, Mode=OneWay, UpdateSourceTrigger=Explicit}"/>
+                <nodes1:DynamoTextBox x:Name="StepTb" Grid.Row="2" Grid.Column="1" Margin="2"
+                                        VerticalContentAlignment="Center" 
+                                        Text="{Binding StepText, Mode=OneWay, UpdateSourceTrigger=Explicit}"/>
+            </Grid>
+            
+        </Expander>
+
+        <Grid Height="{x:Static ui2:Configurations.PortHeightInPixels}" 
+              VerticalAlignment="Top">
+
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="30"/>
+                <ColumnDefinition Width="100"/>
+                <ColumnDefinition MinWidth="150" Width="150"/>
+            </Grid.ColumnDefinitions>
+
+            <nodes1:DynamoTextBox x:Name="ValTb" 
+                                          Text="{Binding ValueText, Mode=OneWay}" 
+                                          Grid.Column="1"
+                                          Margin="2,0,2,0" 
+                                          VerticalAlignment="Stretch"/>
+
+            <Slider Name="slider" 
+                            VerticalAlignment="Center"
+                            Maximum="{Binding Max}" 
+                            Minimum="{Binding Min}" 
+                            TickFrequency="{Binding Step}" 
+                            Value="{Binding Value}" 
+                            TickPlacement="None" 
+                            IsSnapToTickEnabled="true" 
+                            Grid.Column="2" />
+
+        </Grid>
+        
+    </Grid>
+</UserControl>
